Add support for Mesmerize and Showcase.
This one is pretty easy, as the device is almost identical to the Fasciante
I even set it up to use the Fascinate Kernel to minimize commits adn time
when updating for all three devices.

Please fork the the device trees from here.

https://github.com/AOKPSGS

Use the ics branch, as it is set up for AOKP, and make sure showcase's Dir is name
showcasemtd, I named it showcase when I created it by accident.

Change-Id: I5c4989a507ef98f5355c60672d1de3730fd3e25b